Add a migrat service type switch page
